requirements.txt
ipykernel
python-dotenv
numpy==1.26.4
pandas==2.2.3
matplotlib==3.10.1
seaborn==0.13.2
streamlit==1.42.2
scikit-learn==1.6.1
nltk==3.9.1
gensim==4.3.3
torch==2.6.0
tensorflow==2.19.0
tf-keras
torchvision==0.21.0
langchain==0.3.21
langsmith==0.3.19
langserve==0.3.1
langchain-text-splitters==0.3.7
langchain-huggingface==0.1.2
langchain-ollama==0.3.0
langchain-chroma==0.2.2
langchain-openai==0.3.9
langchain-community==0.3.20
langchain-groq==0.3.1
pypdf==5.4.0
bs4==0.0.2
arxiv==2.1.3
PyMuPDF==1.25.4
wikipedia==1.4.0
chromadb==0.6.3
faiss-cpu==1.10.0
sentence-transformers==4.0.1
fastapi==0.115.12
uvicorn==0.34.0
sse-starlette==0.10.0
duckduckgo-search==2025.4.4
jax==0.5.3
mysql-connector-python==9.2.0
validators==0.34.0
youtube-transcript-api==1.0.3
datasets==3.5.0
evaluate==0.4.3
optimum==1.24.0
onnx==1.17.0
unstructured==0.17.2
pytube==15.0.0
yt-dlp==2025.3.31
numexpr==2.10.2
huggingface-hub==0.30.2
hf-xet==1.0.3
cassio==0.1.10
tiktoken==0.9.0
PyPDF2==3.0.1
gradio==5.25.0
langchain-nvidia-ai-endpoints==0.3.9
crewai
crewai-tools